+++ src/authn/ccloudProvider.test.ts
@@ -10,13 +10,17 @@ import {
 import { getTestExtensionContext, getTestStorageManager } from "../../tests/unit/testUtils";
 import { Connection } from "../clients/sidecar";
 import { CCLOUD_AUTH_CALLBACK_URI, CCLOUD_CONNECTION_ID } from "../constants";
+import { ccloudAuthSessionInvalidated } from "../emitters";
 import { getSidecar } from "../sidecar";
 import * as ccloud from "../sidecar/connections/ccloud";
 import * as watcher from "../sidecar/connections/watcher";
 import { getStorageManager, StorageManager } from "../storage";
 import { SecretStorageKeys } from "../storage/constants";
+import { getResourceManager } from "../storage/resourceManager";
 import { getUriHandler, UriEventHandler } from "../uriHandler";
 import { ConfluentCloudAuthProvider, getAuthProvider } from "./ccloudProvider";
+import { CCLOUD_SIGN_IN_BUTTON_LABEL } from "./constants";
+import { AuthCallbackEvent } from "./types";
 
 const TEST_CCLOUD_AUTH_SESSION: vscode.AuthenticationSession = {
   id: TEST_CCLOUD_CONNECTION.id!,
@@ -28,11 +32,14 @@ const TEST_CCLOUD_AUTH_SESSION: vscode.AuthenticationSession = {
   scopes: [],
 };
 
-describe("ConfluentCloudAuthProvider", () => {
+describe("authn/ccloudProvider.ts ConfluentCloudAuthProvider methods", () => {
   let authProvider: ConfluentCloudAuthProvider;
   let uriHandler: UriEventHandler;
 
   let sandbox: sinon.SinonSandbox;
+  // vscode stubs
+  let showErrorMessageStub: sinon.SinonStub;
+  let showInfoMessageStub: sinon.SinonStub;
   // helper function stubs
   let getCCloudConnectionStub: sinon.SinonStub;
   let createCCloudConnectionStub: sinon.SinonStub;
@@ -50,9 +57,8 @@ describe("ConfluentCloudAuthProvider", () => {
   });
 
   beforeEach(() => {
-    authProvider = getAuthProvider();
-
     sandbox = sinon.createSandbox();
+
     getCCloudConnectionStub = sandbox.stub(ccloud, "getCCloudConnection");
     createCCloudConnectionStub = sandbox.stub(ccloud, "createCCloudConnection");
     deleteConnectionStub = sandbox.stub(ccloud, "deleteCCloudConnection");
@@ -62,10 +68,14 @@ describe("ConfluentCloudAuthProvider", () => {
       .stub(watcher, "waitForConnectionToBeStable")
       .resolves(TEST_AUTHENTICATED_CCLOUD_CONNECTION);
 
+    authProvider = getAuthProvider();
     // don't handle the progress notification, openExternal, etc in this test suite
     browserAuthFlowStub = sandbox.stub(authProvider, "browserAuthFlow").resolves();
     stubOnDidChangeSessions = sandbox.createStubInstance(vscode.EventEmitter);
     authProvider["_onDidChangeSessions"] = stubOnDidChangeSessions;
+
+    showErrorMessageStub = sandbox.stub(vscode.window, "showErrorMessage").resolves();
+    showInfoMessageStub = sandbox.stub(vscode.window, "showInformationMessage").resolves();
   });
 
   afterEach(() => {
@@ -80,7 +90,7 @@ describe("ConfluentCloudAuthProvider", () => {
     createCCloudConnectionStub.resolves(TEST_AUTHENTICATED_CCLOUD_CONNECTION);
     getCCloudConnectionStub.onSecondCall().resolves(TEST_AUTHENTICATED_CCLOUD_CONNECTION);
     // authentication completes successfully
-    browserAuthFlowStub.resolves(true);
+    browserAuthFlowStub.resolves({ success: true, resetPassword: false });
 
     await authProvider.createSession();
 
@@ -91,7 +101,7 @@ describe("ConfluentCloudAuthProvider", () => {
   it("createSession() should reuse an existing CCloud connection", async () => {
     getCCloudConnectionStub.resolves(TEST_AUTHENTICATED_CCLOUD_CONNECTION);
     // authentication completes successfully
-    browserAuthFlowStub.resolves(true);
+    browserAuthFlowStub.resolves({ success: true, resetPassword: false });
 
     await authProvider.createSession();
 
@@ -103,7 +113,7 @@ describe("ConfluentCloudAuthProvider", () => {
     const setSecretStub = sandbox.stub(getStorageManager(), "setSecret").resolves();
     getCCloudConnectionStub.resolves(TEST_AUTHENTICATED_CCLOUD_CONNECTION);
     // authentication completes successfully
-    browserAuthFlowStub.resolves(true);
+    browserAuthFlowStub.resolves({ success: true, resetPassword: false });
 
     await authProvider.createSession();
 
@@ -112,6 +122,49 @@ describe("ConfluentCloudAuthProvider", () => {
       SecretStorageKeys.CCLOUD_AUTH_STATUS,
       TEST_AUTHENTICATED_CCLOUD_CONNECTION.status.authentication.status,
     );
+  });
+
+  it("createSession() should handle authentication failure", async () => {
+    getCCloudConnectionStub.resolves(TEST_AUTHENTICATED_CCLOUD_CONNECTION);
+    // authentication fails
+    browserAuthFlowStub.resolves({ success: false, resetPassword: false });
+
+    const authFailedMsg = "Confluent Cloud authentication failed. See browser for details.";
+    await assert.rejects(authProvider.createSession(), {
+      message: authFailedMsg,
+    });
+
+    sinon.assert.calledWith(showErrorMessageStub, authFailedMsg);
+  });
+
+  it("createSession() should handle password reset scenario", async () => {
+    getCCloudConnectionStub.resolves(TEST_AUTHENTICATED_CCLOUD_CONNECTION);
+    // password reset occurred
+    browserAuthFlowStub.resolves({ success: false, resetPassword: true });
+
+    await assert.rejects(authProvider.createSession(), {
+      message: "User reset their password.",
+    });
+
+    sinon.assert.calledWith(
+      showInfoMessageStub,
+      "Your password has been reset. Please sign in again to Confluent Cloud.",
+      sinon.match(CCLOUD_SIGN_IN_BUTTON_LABEL),
+    );
+  });
+
+  it("createSession() should handle user cancellation", async () => {
+    getCCloudConnectionStub.resolves(TEST_AUTHENTICATED_CCLOUD_CONNECTION);
+    // user cancels the operation
+    browserAuthFlowStub.resolves(undefined);
+
+    await assert.rejects(authProvider.createSession(), {
+      message: "User cancelled the authentication flow.",
+    });
+
+    sinon.assert.notCalled(deleteConnectionStub);
+    sinon.assert.notCalled(showInfoMessageStub);
+    sinon.assert.notCalled(showErrorMessageStub);
   });
 
   it("getSessions() should treat connections with a NO_TOKEN/FAILED auth status as nonexistent", async () => {
@@ -204,6 +257,7 @@ describe("ConfluentCloudAuthProvider", () => {
     assert.strictEqual(authProvider["_session"], null);
     assert.ok(deleteSecretStub.calledWith(SecretStorageKeys.AUTH_SESSION_EXISTS));
     assert.ok(deleteSecretStub.calledWith(SecretStorageKeys.AUTH_COMPLETED));
+    assert.ok(deleteSecretStub.calledWith(SecretStorageKeys.AUTH_PASSWORD_RESET));
     assert.ok(stubOnDidChangeSessions.fire.called);
     assert.ok(
       stubOnDidChangeSessions.fire.calledWith({
@@ -244,15 +298,134 @@ describe("ConfluentCloudAuthProvider", () => {
 
   for (const success of [true, false] as const) {
     it(`should return '${success}' from waitForUriHandling when the URI query contains 'success=${success}'`, async () => {
-      const promise: Promise<boolean> = authProvider.waitForUriHandling();
+      const promise: Promise<AuthCallbackEvent> = authProvider.waitForUriHandling();
 
       const uri = vscode.Uri.parse(CCLOUD_AUTH_CALLBACK_URI).with({ query: `success=${success}` });
       uriHandler.handleUri(uri);
 
-      const result: boolean = await promise;
-      assert.strictEqual(result, success);
+      const result: AuthCallbackEvent = await promise;
+      assert.strictEqual(result.success, success);
     });
   }
+});
+
+describe("authn/ccloudProvider.ts ConfluentCloudAuthProvider URI handling", () => {
+  let authProvider: ConfluentCloudAuthProvider;
+
+  let sandbox: sinon.SinonSandbox;
+  // vscode stubs
+  let showInfoMessageStub: sinon.SinonStub;
+  // helper function stubs
+  let deleteConnectionStub: sinon.SinonStub;
+  // auth provider stubs
+  let createSessionStub: sinon.SinonStub;
+  let stubOnDidChangeSessions: sinon.SinonStubbedInstance<
+    vscode.EventEmitter<vscode.AuthenticationProviderAuthenticationSessionsChangeEvent>
+  >;
+
+  before(async () => {
+    await getTestExtensionContext();
+  });
+
+  beforeEach(() => {
+    sandbox = sinon.createSandbox();
+
+    showInfoMessageStub = sandbox.stub(vscode.window, "showInformationMessage").resolves();
+
+    deleteConnectionStub = sandbox.stub(ccloud, "deleteCCloudConnection");
+
+    authProvider = getAuthProvider();
+    createSessionStub = sandbox.stub(authProvider, "createSession").resolves();
+    // don't handle the progress notification, openExternal, etc in this test suite
+    stubOnDidChangeSessions = sandbox.createStubInstance(vscode.EventEmitter);
+    authProvider["_onDidChangeSessions"] = stubOnDidChangeSessions;
+
+    // assume the connection is immediately usable for most tests
+    sandbox
+      .stub(watcher, "waitForConnectionToBeStable")
+      .resolves(TEST_AUTHENTICATED_CCLOUD_CONNECTION);
+  });
+
+  afterEach(() => {
+    // reset the singleton instance between tests
+    ConfluentCloudAuthProvider["instance"] = null;
+    sandbox.restore();
+  });
+
+  it("showResetPasswordNotification() should display a message with sign-in button", async () => {
+    // user dismissed the notification
+    showInfoMessageStub.resolves(undefined);
+
+    authProvider.showResetPasswordNotification();
+
+    sinon.assert.calledWith(
+      showInfoMessageStub,
+      "Your password has been reset. Please sign in again to Confluent Cloud.",
+      CCLOUD_SIGN_IN_BUTTON_LABEL,
+    );
+    sinon.assert.notCalled(createSessionStub);
+  });
+
+  it("showResetPasswordNotification() should call createSession when the sign-in button is clicked", async () => {
+    const clock = sandbox.useFakeTimers(Date.now());
+    // user clicked the sign-in button
+    showInfoMessageStub.resolves(CCLOUD_SIGN_IN_BUTTON_LABEL);
+
+    authProvider.showResetPasswordNotification();
+    // simulate the passage of time to allow the notification to be shown + button clicked
+    await clock.tickAsync(100);
+
+    sinon.assert.calledWith(
+      showInfoMessageStub,
+      "Your password has been reset. Please sign in again to Confluent Cloud.",
+      CCLOUD_SIGN_IN_BUTTON_LABEL,
+    );
+    sinon.assert.calledOnce(createSessionStub);
+  });
+
+  for (const success of [true, false] as const) {
+    it(`handleUri() should not invalidate the current CCloud auth session for non-reset-password URI callbacks (success=${success})`, async () => {
+      const setAuthFlowCompletedStub = sandbox
+        .stub(getResourceManager(), "setAuthFlowCompleted")
+        .resolves();
+      const deleteSecretStub = sandbox.stub(getStorageManager(), "deleteSecret").resolves();
+      const showResetPasswordNotificationStub = sandbox.stub(
+        authProvider,
+        "showResetPasswordNotification",
+      );
+
+      const uri = vscode.Uri.parse(CCLOUD_AUTH_CALLBACK_URI).with({ query: `success=${success}` });
+      await authProvider.handleUri(uri);
+
+      sinon.assert.calledWith(setAuthFlowCompletedStub, { success, resetPassword: false });
+      sinon.assert.notCalled(deleteConnectionStub);
+      sinon.assert.notCalled(deleteSecretStub);
+      sinon.assert.notCalled(showResetPasswordNotificationStub);
+    });
+  }
+
+  it("handleUri() should invalidate the current CCloud auth session for reset-password URI callbacks", async () => {
+    const setAuthFlowCompletedStub = sandbox
+      .stub(getResourceManager(), "setAuthFlowCompleted")
+      .resolves();
+    const ccloudAuthSessionInvalidatedFireStub = sandbox.stub(ccloudAuthSessionInvalidated, "fire");
+    const deleteSecretStub = sandbox.stub(getStorageManager(), "deleteSecret").resolves();
+    const showResetPasswordNotificationStub = sandbox.stub(
+      authProvider,
+      "showResetPasswordNotification",
+    );
+
+    const uri = vscode.Uri.parse(CCLOUD_AUTH_CALLBACK_URI).with({
+      query: "success=false&reset_password=true",
+    });
+    await authProvider.handleUri(uri);
+
+    sinon.assert.calledWith(setAuthFlowCompletedStub, { success: false, resetPassword: true });
+    sinon.assert.calledOnce(deleteConnectionStub);
+    sinon.assert.calledWith(deleteSecretStub, SecretStorageKeys.CCLOUD_AUTH_STATUS);
+    sinon.assert.calledOnce(ccloudAuthSessionInvalidatedFireStub);
+    sinon.assert.calledOnce(showResetPasswordNotificationStub);
+  });
 });
 
 describe("CCloud auth flow", () => {