+++ src/models/environment.ts
@@ -14,7 +14,7 @@ import {
   LOCAL_ENVIRONMENT_NAME,
   UTM_SOURCE_VSCODE,
 } from "../constants";
-import { FormConnectionType } from "../webview/direct-connect-form";
+import { FormConnectionType } from "../directConnections/types";
 import {
   CCloudKafkaCluster,
   DirectKafkaCluster,
@@ -231,7 +231,7 @@ export class EnvironmentTreeItem extends TreeItem {
     } else if (isDirect(resource)) {
       const { missingKafka, missingSR } = checkForMissingResources(resource);
       if (missingKafka || missingSR) {
-        this.iconPath = new ThemeIcon("error", new ThemeColor("problemsErrorIcon.foreground"));
+        this.iconPath = new ThemeIcon("warning", new ThemeColor("problemsErrorIcon.foreground"));
       }
     }
     this.tooltip = createEnvironmentTooltip(this.resource);